unify account switcher implementation
